Added pch to projects

+++ projects/core/Core/sources/Application.cpp
@@ -1,12 +1,7 @@
-#include "GraphicsCore.h"
-
-#include <glm/gtc/matrix_transform.hpp>
+#include "pch.h"
 
 #include "Application.h"
-
-#include "SimulationRunner.h"
 #include "Layers.h"
-#include "Logger.h"
 
 #include <string>
 #include <vector>
@@ -49,7 +44,7 @@ void Application::run()
 
     const auto& guiWindow   = *guiWindows.get<NotificationWindow>("test");
     const auto& debugWindow = *guiWindows.get<DebugWindow>("debug");
-    auto& demoWindow        = *guiWindows.get<DemoWindow>("demo");
+          auto& demoWindow  = *guiWindows.get<DemoWindow>("demo");
     const auto& propWindow  = *guiWindows.get<PropulsionControlWindow>("prop");
     // ***********
 
@@ -77,7 +72,7 @@ void Application::run()
         centralRenderer.clear();
 
         float scale = 0.05f * debugWindow.scale;
-        glm::mat4 proj = glm::ortho(-1.0f * scale * window.getWidth(), 1.0f * scale * window.getWidth(), -1.0f * scale * window.getWidth(), 1.0f * scale * window.getHeight(), -1.0f, 1.0f);
+        glm::mat4 proj = glm::ortho(-1.0f * scale * (float)window.getWidth(), 1.0f * scale * (float)window.getWidth(), -1.0f * scale * (float)window.getWidth(), 1.0f * scale * (float)window.getHeight(), -1.0f, 1.0f);
         glm::mat4 view = glm::translate(glm::mat4(1.0f), glm::vec3(-player.position[0], -player.position[1], 0));
         if (!propWindow.followPlayer)
         {
